transactions: some more compaction improvements

+++ tests/java/e2e-verifiers/src/main/java/org/apache/kafka/tools/VerifiableProducer.java
@@ -202,7 +202,7 @@ public class VerifiableProducer implements AutoCloseable {
   // Send messages with a key of 0 incrementing by 1 for
   // each message produced when number specified is reached
   // key is reset to 0
-  private final Integer repeatingKeys;
+  private final Long repeatingKeys;
 
   private int keyCounter;
 
@@ -217,7 +217,7 @@ public class VerifiableProducer implements AutoCloseable {
 
   public VerifiableProducer(
       Properties properties, String topic, int throughput, int maxMessages,
-      Integer valuePrefix, Long createTime, Integer repeatingKeys,
+      Integer valuePrefix, Long createTime, Long repeatingKeys,
       boolean isTransactional, boolean enableRandomAborts, long batchSize) {
 
     this.topic = topic;
@@ -332,7 +332,7 @@ public class VerifiableProducer implements AutoCloseable {
     parser.addArgument("--repeating-keys")
         .action(store())
         .required(false)
-        .type(Integer.class)
+        .type(Long.class)
         .metavar("REPEATING-KEYS")
         .dest("repeatingKeys")
         .help(
@@ -420,7 +420,7 @@ public class VerifiableProducer implements AutoCloseable {
     String configFile = res.getString("producer.config");
     Integer valuePrefix = res.getInt("valuePrefix");
     Long createTime = res.getLong("createTime");
-    Integer repeatingKeys = res.getInt("repeatingKeys");
+    Long repeatingKeys = res.getLong("repeatingKeys");
 
     if (createTime == -1L) createTime = null;
